+++ lib/plausible/application.ex
@@ -4,6 +4,10 @@ defmodule Plausible.Application do
   use Application
   use Plausible
 
+  on_ee do
+    @start_verification_mock_scenarios? Mix.env() in [:dev, :e2e_test, :test]
+  end
+
   def start(_type, _args) do
     on_ee(do: Plausible.License.ensure_valid_license())
     on_ce(do: :inet_db.set_tcp_module(:happy_tcp))
@@ -194,6 +198,11 @@ defmodule Plausible.Application do
         end,
         Plausible.Ingestion.Counters,
         Plausible.Session.Salts,
+        on_ee do
+          if always(@start_verification_mock_scenarios?) do
+            Plausible.InstallationSupport.Verification.MockScenarios
+          end
+        end,
         Supervisor.child_spec(Plausible.Event.WriteBuffer, id: Plausible.Event.WriteBuffer),
         Supervisor.child_spec(Plausible.Session.WriteBuffer, id: Plausible.Session.WriteBuffer),
         ReferrerBlocklist,