+++ lib/plausible_web/controllers/site_controller.ex
@@ -50,11 +50,7 @@ defmodule PlausibleWeb.SiteController do
         end
 
         redirect(conn,
-          to:
-            Routes.site_path(conn, :installation, site.domain,
-              site_created: true,
-              flow: flow
-            )
+          to: Routes.site_path(conn, :installation, site.domain, flow: flow)
         )
 
       {:error, _, :permission_denied, _} ->
@@ -169,8 +165,14 @@ defmodule PlausibleWeb.SiteController do
     )
   end
 
-  def settings_email_reports(conn, _params) do
+  def settings_email_reports(conn, params) do
     site = conn.assigns[:site]
+
+    if params["cta_clicked"] == "true" do
+      site
+      |> Plausible.Site.put_onboarding_status_advance(:completed)
+      |> Repo.update!()
+    end
 
     conn
     |> render("settings_email_reports.html",